renames and generalizes AutoScroll function a little more. generall code cleanup
