+++ src/components/navigationButtons/internalNavButton.js
@@ -1,6 +1,6 @@
 import React from "react";
 import {Link} from "react-router-dom";
-import ScrollToTop from "../scrollToTop/scrollToTop";
+import AutoScroll from "../autoScroll/autoScroll";
 import "./navButton.scss";
 
 /**
@@ -18,7 +18,7 @@ export default function InternalNavButton({pathName, buttonText}) {
 			to={pathName}
 			style={{textDecoration: "none"}}
 		>
-			<ScrollToTop/>
+			<AutoScroll/>
 			{buttonText}
 		</Link>
 	);
